+++ src/Console/TenantRouteListCommand.php
@@ -20,7 +20,6 @@ namespace Somnambulist\Tenancy\Console;
 
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
-use Illuminate\Http\Request;
 use Illuminate\Routing\Route;
 use Illuminate\Routing\Controller;
 use Symfony\Component\Console\Input\InputOption;
@@ -99,7 +98,7 @@ class TenantRouteListCommand extends AbstractTenantCommand
         }
 
         if ($sort = $this->option('sort')) {
-            $results = array_sort(
+            $results = Arr::sort(
                 $results,
                 function ($value) use ($sort) {
                     return $value[$sort];
@@ -157,10 +156,6 @@ class TenantRouteListCommand extends AbstractTenantCommand
     protected function getMiddleware($route)
     {
         $middlewares = array_values($route->middleware());
-
-        $middlewares = array_unique(
-            array_merge($middlewares, $this->getPatternFilters($route))
-        );
 
         $actionName = $route->getActionName();
 
@@ -228,44 +223,6 @@ class TenantRouteListCommand extends AbstractTenantCommand
     }
 
     /**
-     * Get all of the pattern filters matching the route.
-     *
-     * @param  \Illuminate\Routing\Route $route
-     *
-     * @return array
-     */
-    protected function getPatternFilters($route)
-    {
-        $patterns = [];
-
-        foreach ($route->methods() as $method) {
-            // For each method supported by the route we will need to gather up the patterned
-            // filters for that method. We will then merge these in with the other filters
-            // we have already gathered up then return them back out to these consumers.
-            $inner = $this->getMethodPatterns($route->uri(), $method);
-
-            $patterns = array_merge($patterns, array_keys($inner));
-        }
-
-        return $patterns;
-    }
-
-    /**
-     * Get the pattern filters for a given URI and method.
-     *
-     * @param  string $uri
-     * @param  string $method
-     *
-     * @return array
-     */
-    protected function getMethodPatterns($uri, $method)
-    {
-        return $this->router->findPatternFilters(
-            Request::create($uri, $method)
-        );
-    }
-
-    /**
      * Filter the route by URI and / or name.
      *
      * @param  array $route
@@ -274,7 +231,8 @@ class TenantRouteListCommand extends AbstractTenantCommand
      */
     protected function filterRoute(array $route)
     {
-        if (($this->option('name') && !Str::contains($route['name'], $this->option('name'))) ||
+        if (
+            ($this->option('name') && !Str::contains($route['name'], $this->option('name'))) ||
             $this->option('path') && !Str::contains($route['uri'], $this->option('path')) ||
             $this->option('method') && !Str::contains($route['method'], $this->option('method'))
         ) {
@@ -296,13 +254,7 @@ class TenantRouteListCommand extends AbstractTenantCommand
             ['name', null, InputOption::VALUE_OPTIONAL, 'Filter the routes by name.'],
             ['path', null, InputOption::VALUE_OPTIONAL, 'Filter the routes by path.'],
             ['reverse', 'r', InputOption::VALUE_NONE, 'Reverse the ordering of the routes.'],
-            [
-                'sort',
-                null,
-                InputOption::VALUE_OPTIONAL,
-                'The column (host, method, uri, name, action, middleware) to sort by.',
-                'uri'
-            ],
+            ['sort', null, InputOption::VALUE_OPTIONAL, 'The column (host, method, uri, name, action, middleware) to sort by.', 'uri'],
         ];
     }
 }